Moved logger related methods to an extension.
